api/requirements.txt
aniso8601==8.1.0
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
google-cloud-bigquery==1.27.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.800
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.8
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-flask==1.1.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
